requirements.txt
pandas>=1.3.0
nltk==3.8.1
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
torch>=1.9.0
transformers>=4.11.0
beautifulsoup4>=4.12.2
streamlit>=1.31.0
groq>=0.9.0
python-dotenv>=1.0.0
